Changes to WebClientConfig @ added retrieve and image for a vehicle.

+++ Server/src/main/java/com/capstone/models/VehicleInfo.java
@@ -46,7 +46,7 @@ public class VehicleInfo {
         return vehicleInfoId;
     }
 
-    public void setVehicleInfoId(int vehicleInfoId) {
+    public void setVehicleInfoId(long vehicleInfoId) {
         this.vehicleInfoId = vehicleInfoId;
     }
